Adds ToStringRenderer that uses Object#toString
This adds ToStringRenderer as a Renderer implementation that uses
toString method of the object. This is useful for Strings, boxing
types and any other object that has a toString implementation
suitable for display.

With ToStringRenderer in place, this patch also adds default ctrs
to various Renderer based widgets. (Acutally relying on toString
is the default behavior in all Swing widgets).

Change-Id: If9ea40db1e83f549436fc540f5d0c6a264703a04