Dramatically improves generics support in UiBinder.
This is actually based on my previous patch for UiHandler to
enchance type information based on @UiField declarations.
This moves type enhancing to an earlier step so that we get
more info on the type for all fields.

This added complication to the other parts of the UiBinder
that dealed with generics by simply 'ignoring' them so I
needed to fix them as well.

Change-Id: I25bb0488c6ea3816d9b34bdf08cd0f5906bda5ab
Review-Link: https://gwt-review.googlesource.com/#/c/5842/