Update patch set 1
Patch Set 1: -Code-Review

Opps, I didn't meant +2, reverting...

The main problem is, you cannot blindly call addHandler as that wouldn't work with dom events.
You can take a look at the event type and choose the method based on that but I'm not sure if this feature would worth all the trouble.

Patch-set: 1