Update patch set 2
Patch Set 2: Code-Review-2

Browser are exposing sub pixel values (in rendering and events) even more. 

The GWT DOM APIs will need a way to deal with those values. The rounding and the private methods were only introduced for people that are not willing to deal with sub pixel rendering, but sooner or later we will need to expose sub pixel values as well.

Making those methods private was only done to delay the discussion if this is the best way to do it.

If we care about optimizing code size (or a bit runtime performance here) we should add a compiler pass that changes (value || 0) | 0 to value |0 instead.

Patch-set: 2
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-2