Update patch set 2
Patch Set 2:

In general, we should avoid adding & keeping a code just for future discussion. This leaves the code in half baked state for undefined length of time without adding much value; and mostly confuses the reader.

This is kind of changes are quite easy to introduce once/if a decision is made. I'm not thinking strongly that we should submit this specific change however we should not get in to this state in the first place.

 > Browser are exposing sub pixel values (in rendering and events)
 > even more.
 > 
 > The GWT DOM APIs will need a way to deal with those values. The
 > rounding and the private methods were only introduced for people
 > that are not willing to deal with sub pixel rendering, but sooner
 > or later we will need to expose sub pixel values as well.
 > 
 > Making those methods private was only done to delay the discussion
 > if this is the best way to do it.
 > 
 > If we care about optimizing code size (or a bit runtime performance
 > here) we should add a compiler pass that changes (value || 0) | 0
 > to value |0 instead.

Patch-set: 2