Update patch set 1
Patch Set 1: Code-Review-1

Correct me if I am wrong, but I think there is a problem with this patch:

First of all I agree we need to do something about the dialog since the alert is not helpful at all in prod, but we should be more deterministic about what is happening.

Entrypoints are going to be executed in random order, so there might be an UCE already or not depending on the order of execution. 
reportUncaughtException will report the exception to the UCE (if there is one) or let it pop up in the console of the browser.

Since it does not use the entry function, the exception will not be seen by an UCE that gets registered later, but I think we should be consistent here and not rely on order. And if there is going to be an UCE for this app the exception should arrive there, so it will end up in the apps server side logging.

Patch-set: 1
Label: Code-Review=-1