Update patch set 1
Patch Set 1: -Code-Review

Good point. Perhaps use Scheduler.scheduleFinally?

However, when running code on the wrong browser, things may be so messed up that we don't get through all the entry points. It seems difficult to do server-side logging in that case, but perhaps call GWT.debugger() and log to the console as well, so it has a chance to be noticed in interactive testing?

Patch-set: 1