Update patch set 2
Patch Set 2:

> Could you explain more what this actually does?

Current uibinder implementation is ignoring the generics and remove any type information if there is some.
This is partly because ui.xml cannot declare the type parameters for generic types. However, if the element is declared as a ui:field then we might know more about the type based on the declaration in the class.

I did a similar enhancement before just for @UIHandler but other parts (@UiFactory & field references) are affected with this as well; which this patch fixes. Roberto reviewed the earlier one and already familiar with the problem.

 > For users, what works now that didn't work before?

See the changes in changes c/g/gwt/uibinder/client/test/.

 > Also, are we
 > stricter about type-checking in a way that could cause breakage?

Yes, this is potentially a breaking change though it breaks in compile time and fix is simple.

Patch-set: 2
Label: Verified=0