Update patch set 1
Patch Set 1:

(4 comments)

I disagree with Brian about the prefixes (except for "ms"), because mobile browsers need them (Firefox for Android, and Safari Mobile on iOS 6).

Re. cancelAnimationFrame, could we do both? i.e. use a boolean flag in the AnimationHandle *and* use cancelAnimationFrame if it exists or cancelTimeout if we polyfilled requestAnimationFrame to use a timer.

I.e.

* if we don't have requestAnimationFrame, polyfill requestFunction and cancelFunction to use setTimeout and cancelTimeout, store the timer ID in the AnimationHandle.
* otherwise, if we have cancelAnimationFrame, set the cancelFunction to it; store the ID returned by requestAnimationFrame in the AnimationHandle
* otherwise, leave cancelFunction to null.

On cancel, if cancelFunction is not null, call it with the ID stored in the AnimationHandle; otherwise just set the flag in the AnimationHandle.

When the event (or timer) fires, always check for the boolean flag in in AnimationHandle.

Patch-set: 1
Label: Verified=0