Update patch set 5
Patch Set 5:

I uploaded a new version that adds non-prefixed API support and reuses the old Timer implementation.

One point that wasn't still clear to me is if we need to keep support for prefixed APIs or not. If we don't need than it makes sense to get rid of rebindings. However that means it will be harder to patch from outside as we no longer use GWT.create.

If we need to support the prefixed versions than we can revive the old code and put back the rebindings.

Patch-set: 5
Label: Verified=0