Update patch set 7
Patch Set 7:

I agree it's not a perfect situation with the hibernate validator issue. One possibility that I see is to not export it and have an additional bundle fragment which exports the troublesome package. But that's definitely for later.

I agree with your assessment and will not hold back the review for just this issue. Let me know if you want me to test some additional changes, or just this state of the patch set and I'll try to expedite.

Patch-set: 7