Update patch set 3
Patch Set 3:

I tried it with Showcase, and while the code runs fine, the JavaScript for class method initialization looks a bit unfortunate. Here's an example:

_.categoryPanels = cggssc.categoryPanels = function(){
  return 'Panels';
}

I believe it's because it originally looked something like this:

_.categoryPanels = function categoryPanels() {
  return 'Panels';
}

After rewriting the function, the result is a double-assignment.

In the Chrome debugger's call stack, it displays both names, which is a bit confusing.

The names added to the package namespace seem to be otherwise unused in most cases. I'm wondering if we should do dead global removal, or perhaps rewrite the double-assignment as two assignments?

Patch-set: 3
Label: Verified=0