Update patch set 4
Patch Set 4:

i've uploaded patch set 4 which addresses all of your comments (except for the request for tests for RuleGenerateWith) and does a lot of work to modularize JavaToJavaScriptCompiler in the way we discussed.

there's some remaining cleanup work in JavaToJavaScriptCompiler, i haven't yet turned all of the stages in precompile() and compilePerms() into functions. the code is really messy and it's not yet clear to me how to group the stages in a way that isn't purely arbitrary

Patch-set: 4
Label: Verified=0