Update patch set 1
Patch Set 1:

Sorry for all that. It's the first time I submit a patch and I am not familiar with this code-review process.

I did not associate this patch to the bug since I added the link to this patch in the issue tracker, so I had to commit it before.

I do not know how to send the patch to the 2.6 release. I assume I have to use the cherry pick feature but as I am not sure about that, I would rather like you to handle it or explain me how I can do that (for the next time).

Patch-set: 1