Speed-up ArtifactSet.
This introduces a Multimap, which makes selection of matching entries O(# of 
Artifact subclasses) instead of O(# of Artifacts).

Additionally, this change caches results of finds.  Experimentally, this seems 
to make a difference.

Downsides: It consumes more memory to keep the Multiset, the results of find are 
unmodifiable, and it's slower for Sets where find is not called.

Change-Id: I8411f44216f14f9bc0eb7b2a0860343d23babfc2