Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > > > Can't you do that by implementing your SuggestionOracle
 > > > accordingly?
 > >
 > > You're still incurring expensive validation computations on every
 > > keystroke.  Also, that wouldn't solve the problem of the user
 > > pressing the button without having changed the query, which is
 > > suppressed by the SuggestBox and never gets to the oracle.
 > 
 > Query being changed or not should be an issue, you can easily check
 > that.
 > 
 > I'm not sure what you are referring to with expensive validations
 > computations?
 > 
 > Also I'm not that much familiar with the code. I'll bring in
 > tbroyer to chime in as well.

The SuggestBox checks whether the current query is the same as the last query it made.  If it is, it does not call out to the oracle.  Because of this, I don't think it's possible to get my desired interaction with only changes at the Oracle level.

Patch-set: 1