Update patch set 1
Patch Set 1: Code-Review-1

> Generating suggestions requires some expensive computation, so I
 > want to do it only when the user clicks a button.

You'd better not use a SuggestBox then. What you're describing is a TextBox+Button, and on click of the Button you show a PopupPanel with a list of things (using Menu?). That doesn't rule out using SuggestOracle or even SuggestionDisplay.

But as Goktug said, even with SuggestBox I don't think you need the changes you're proposing here: use a SuggestOracle that can be "enabled" or "disabled". When disabled, it doesn't answer, or answers with an empty list of suggestions; this is used to short-circuit the "key up" (and yes, it incurs a performance loss, but I believe it's negligible in this case). When you click the button, "enable" the SuggestOracle, and call showSuggestionList(); and handle the case where the query is the same as before in your SuggestOracle by keeping a cache of the last request and response.

Anyway, I think you misread showSuggestionList() (and I hope I didn't myself): it clears currentText, so the check in refreshSuggestions is always false, which means showSuggestions is always called in this case, whether the query has changed from the last time you called showSuggestionList() (clicked the button) or not.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1