Update patch set 1
Patch Set 1:

for years now there has been an undocumented requirement that subclasses of SelectionScriptLinker must provide an implementation of __gwt_isKnownPropertyValue and __gwt_getMetaProperty.

i've added the missing documentation of this requirement and also mentioned this new extension of it (that these functions must also be made available to the application JS scope).

ray: i'll take care of updating any internal custom linkers.

goktug: agreed that it would be nice for this to be automatic. if the linker generated nocache.js started as generated Java that was compiled, this would be pretty trivially implemented in a base class. but at it is now as JS there doesn't seem to be a consistent way.

Patch-set: 1
Label: Verified=0