Update patch set 4
Patch Set 4:

So, late on Sunday, I posted the patch set that attempts to keep views as mutable as possible.  It does this by sharing the cache reference between the root set and all of its children views.  When there is a mutate anywhere, the cache is cleared.

Find remains immutable (and was converted to an ImmutableSortedSet).

I also reprofiled, and the sum of all ArtifactSet operations <= 0.1% for CPU time.  For context, the total SuperDevMode compile time for this app takes about 2 minutes.

Patch-set: 4