Update patch set 1
Patch Set 1:

> I'd prefer we remove it only when we remove the rest of
 > c.g.g.event.shared that has a replacement in c.g.w.b.event. Is it
 > causing any harm to leave it here?

It doesn't cause any harm, we can leave it there. It is just a long deprecated code and is not used internally in GWT nor Google.
I think the question is why do you want to keep this code and why do you want a replacement for it? I can't see any use for it?

Patch-set: 1
Label: Verified=0