Update patch set 1
Patch Set 1:

> I'm perfectly fine either way, but just a comment from the peanut
 > gallery, in that I thought the c.g.w.bindery movement had kind of
 > fizzled out with the decreased emphasis on RF/etc.
 > 
 > Do we still have a good vision for what "c.g.web.bindery" is all
 > about? IIRC it was for components that could be cross-GWT/Android/etc.
 > ? Is that still the idea?
 > 
 > Just curious.

I have been asking the similar questions to myself.

Again as many others, this is something left in the middle of refactoring and never completed. Actually I was looking into finishing the refactor of just Activities and remove the old EventBus from there but it will be difficult to achieve in Google without Java 8 defender methods.

There is still a vision for cross-GWT/Android apps but I don't think c.g.web.bindery is part of it. Our hybrid apps are not using it. Also we are not recommending the use of RF for a while. The open-source community looks like moving in the same direction as well.
For GWT, I'm pushing & evangelizing for 2 RPC mechanisms - JAX-RS and gwt protobufs; those are probably what we might want to support in the long run. And hopefully we will be investing towards that.

Going back to the eventbus, at the end, I'm not sure it would be better to:

1- Finish the transformation to get rid of c.g.gwt.event.shared 

2- Revert some changes back and have two distinct EventBus implementations; one for c.g.gwt.event.shared and one for c.g.web.bindery.

3- Leave it alone in its current state.

I haven't analyzed the implications all of the options yet but even though I don't like the 3rd option, that is what is happening right now.

Patch-set: 1