Enable null checks for 'this' in Super Dev Mode
The GWT compiler converts non-polymorphic methods
into static functions, but doesn't normally check that
'this' is non-null. This results in confusing errors
for Java programmers who would expect a
NullPointerException.

In Super Dev Mode we will throw TypeError("null pointer")
for a null 'this'. (However, the null check isn't done for
JavaScriptObjects because tricky code exists that takes
advantage of null checks not being done.)

Change-Id: I49592fa03575d17a7bf1815d3cb66938e6d18d00