Update patch set 1
Patch Set 1: Code-Review+1

I recall this "tricky code" coming up in past discussions about this (when trying to fix this in DevMode IIRC); are we just accepting that we'll never be able to fix this?

E.g. I'm curious what this tricky code actually is, and if there's a time table to somehow have people fix it.

I understand it's not like we can add a deprecation warning...

Patch-set: 1
Reviewer: Gerrit User 5025 <5025@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1