Update patch set 1
Patch Set 1:

I'm doing the simplest thing to get most null checking turned on. We can go back to deal with JavaScriptObject later. (Assuming we get to it.)

I don't know how widespread it is. The first example I ran into was in CellBasedWidgetStandard where it's doing this:

  target = target.getParentElement().cast();

And then checking if target is null.

We might be able to implement it behind a flag, fix all the occurrences in GWT itself, and leave it off by default. Or use an annotation to enable/disable null "this" checking one class at a time?

Patch-set: 1
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0