Update patch set 1
Patch Set 1:

> > > I'd prefer we remove it only when we remove the rest of
 > > > c.g.g.event.shared that has a replacement in c.g.w.b.event. Is
 > it
 > > > causing any harm to leave it here?
 > >
 > > It doesn't cause any harm, we can leave it there. It is just a
 > long
 > > deprecated code and is not used internally in GWT nor Google.
 > > I think the question is why do you want to keep this code and why
 > > do you want a replacement for it? I can't see any use for it?
 > 
 > I just meant to consider c.g.g.event.shared as a whole (modulo
 > GwtEvent and HandlerManager).

I think DefaultHandlerRegistration has nothing to do with the rest of the package right now and it looks like a leftover from an even earlier iteration.
It is not a meaningful default implementation for neither older or newer HandlerRegistration. That just adds more to the confusion.

I think it should be deleted regardless of our plans with c.g.gwt.event or c.g.web.bindery; don't you think so?

Patch-set: 1