Update patch set 1
Patch Set 1: Code-Review+2

W.r.t checking nulls on JSOs devirtual methods, I think we will have to bite the bullet and do it at some time. It it not used that much on purpose, but we have quite few test that fail to mock some objects and don't fail because of null is not checked.

Additionally for interfaces it is really inconsistent, if an interface is only implemented by a JSO then the it will dispatch on null with no problem but if it is implemented both by a JSO and Java objects, it will throw an exception on null dispatch. These two cases should behave the same way.

Patch-set: 1
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2