Update patch set 1
Patch Set 1:

I'm sure it was useful in the early days, but since then we haven't seen any regressions handling multiple modules for all the time that this test hasn't been run. Or at least so we think; are we sure that it fails now due to a spurious failure and not a subtle but real issue?

If it didn't find a real bug then I think we might as well remove it.

Patch-set: 1