Update patch set 4
Patch Set 4:

I decided to revert the extra logic that I have added for vertical alignment. 

A proper implementation should flip to menu to show above or below the parent menu depending on available space - if it still doesn't fit that it can align with top or bottom (in analogy to what I have done for horizontal).

However usually vertical scroll is not much big of a deal and problems occurs less frequently. So I'll favor simplicity for now. If enough people complain about it in the issue tracker, we can re-consider this.

Patch-set: 4
Label: Verified=0