Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Thanks for the patch.

Please fix the import (the generated annotation value is not a blocker).

I also wonder if there shouldn't be a small change to the way we package requestfactory-* JARs, to make sure GwtIncompatible is in those JARs (in case you don't use gwt-serlvet.jar on the server side; or you use RF on a client such as an Android app)

Patch-set: 1
Label: Code-Review=-1