Fix Animation wrt high-res timestamps
Ie00d1020285ce41c571054709bbe19aeef348975 slightly changed the semantics
of the argument to the AnimationCallback, to align it with the
requestAnimationFrame spec: the timestamp might now be a high-res timestamp
so it should only be compared to other similar values. The problem is that
Animation was relying on it being an "absolute" timestamp, so animations
would never complete in browsers using requestAnimationFrame.

Bug: issue 8570
Change-Id: I11dd7daf121ae1898d373d70e31897c78bc41716