Update patch set 1
Patch Set 1:

I wrote a test that should reproduce the issue but I do not manage to execute it.

I followed the eclipse readme file to set-up my workspace but it does not compile and I am not able to run this test from Eclipse. I think I found some mistakes in this document, some little fixes enabled me to go to next steps but I am blocked again.

I tried to run it with ant but I am not able to get it fail, so I assume it is not ran in ant or maybe I don't read the right report.

What is the best way for me to get help regarding this issues?

Thanks in advance.

Patch-set: 1
Reviewer: Gerrit User 6645 <6645@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0