Update patch set 1
Patch Set 1:

> I agree but there's no support for Safari < 5.1 and Android < 3
 > Does GWT support these browsers?

From Google side, I don't think we need to support Safari 5 or Android 2.3 below. Daniel, is that any different from steering committee perspective?

 > 
 > If not, maybe I should create separate package private class
 > ClassList with
 > two implementations?

If it comes to this, I think it would would be better to wait little bit more and get rid of the existing code.

Patch-set: 1