Fixes colliding names from nested classes in RequestFactory.
DeobfuscatorBuilder was ignoring the $ in binary names of nested
classes that were being used as a basis for new DeobfuscatorBuilder
classes.

For example when generating based off of a class with binary name
com.google.Foo$Bar it would create a new class with source name of
com.google.Foo$BarDeobfuscatorBuilder. But that is an invalid source
name for a class which is not actually nested.

As a result the compiled version of these constructed classes ended up
with self reported names of com.google.BarDeobfuscatorBuilder, causing
multiple collisions which were being caught and reported as errors in
TypeOracle.

Change-Id: I16bfaed1a3467f6187b4ca80dcb7b95049e65b50