Update patch set 1
Patch Set 1:

I agree with Andrei, it is a change in style. 

I don't think it is meaningful to evaluate the performance as if it where JavaScript code for two reasons:
- the actual JavaScript output by the compiler might be very different than the Java source code (and if we determine an equivalent faster way to do some operations, then we should make it a compiler optimization); and
- more importantly in this case the relative cost of updating the variable change with respect to the actual call to add() seems negligible to me.

So back to it as a change in style I have a very very slight preference for accepting the change but I defer to you on that.

Patch-set: 1
Label: Verified=0