Update patch set 1
Patch Set 1:

I could go further thanks to the contributor mailing list.

It sounds like Eclipse projects configuration files are not up to date. I fixed them and I am now able to run tests in Eclipse.

But I am not able to reproduce this bug in a GWTUnit test. As this issue depends on browsers, it may not run with the browser that is able to reproduce the issue.

I did a litte project that demonstrates the issue on github: https://github.com/bsautel/gwt-history-encoding-issue

It shows that browsers don't have the same behavior.

Patch-set: 1