Update patch set 1
Patch Set 1:

Yeah, I'm not completely opposed to inherit since it seems potentially useful, but in general it's a lot easier for us to add features than to take them away.  So conservatively I'd suggest leaving it out if you don't have an immediate need, and we can revisit adding it everywhere in a followup CL if appropriate.

Patch-set: 1