Update patch set 10
Patch Set 10:

We don't actually need to roll this back here; if you make another patch, I can import both git patches as a single CL to get the same effect. (It's a bit tricky to upload the patch, though.)

As I said privately, I think delegating all methods in ArtifactSet to TypeIndexedSet is the way to go.

Patch-set: 10