Update patch set 7
Patch Set 7:

Thanks for the link, Cristiano. IMHO it would be best if the OSGi headers would be maintained in GWT, and not in an external project. But that's out of my hands :-)

BTW, it's your change and you should be able to abandon it if you no longer intend to work on it.

Patch-set: 7