Allow to use globally unique strings as type ids.
Stable global id are a necessity for cross module reference when
global information is not available (e.g. separate compilation).

Although not practically necessary, TypeIds in intrinsic are
represented as JavaScriptObjects to denote that the type is
not necessarily int. Leaving them as ints would have been also
functional as AST in the stage where this pass is applied does
not need to be type correct Java.

Change-Id: I05dc27cfe05be1a8367924e4594ddc7619b1ab52