Update patch set 5
Patch Set 5:

> This is how gwt-pushstate works nowadays, but IMO it's a hack.
 > History works with hashes, and provides some guarantees as to the
 > encoding/decoding/roundtripping of the token; if you want something
 > else then use another API. 

IMHO just because we call it pushState does not mean it is a different API. It is still the History API but with added features and the History API is represented by the History class in GWT. Thus I think we should make the pushState methods public on the History class right away and the "classic" hash fragment based methods can then use them if needed. Although it is simple I don't see the point why everyone else should write the same JSNI methods again if they want to use pushState.

Of course this can also be done in a different CL but it would also fit into this one if the commit message will be changed a bit.


 > History is low-level, and you're advised
 > to build something higher-level, such as Places, if you want more
 > freedom as to how your URLs will look like and behave. With Places
 > for example, you can easily plug an Html5Historian:
 > https://gist.github.com/tbroyer/1883821, or something more complex
 > that supports both browsers with and without support for pushState.

Thats totally fine, and it would be even finer if History would have public pushState methods.


 > Hacking into existing deferred-bindings is just that: a hack, *Impl
 > classes are implementation details and subject to change without
 > notice.

Well I agree that its not that nice to replace GWT's *Impls but GWT does not forbid it either. Maybe it should in the future so people do not invest time writing libraries that may replace *Impl classes. Because in case of gwt-pushstate this CL makes this library probably useless which does not feel nice.


I think thats the first time we "disable" a deferred binding by choosing an implementation based on a runtime feature check before hand. So I just wanted to raise some more attention and of course vote for public pushState methods on the History class instead of keeping them private.

Patch-set: 5