Updates module property and resource loading for libraries.
Changed Precompile to not throw an assertion exception during library
compiles that don't have an entry point.

Added differentiation for binding and configuration property
definitions and values that are defined by the current target library
versus defined in the dependency tree. Used this differentiation to
save a record of these values in the current libraryWriter.

Added module parsing support for a <resource> tag as an alternative to
the <source> tag for when needing to pull in build resources (PNG and
CSS) and not wanting to also pull in the java source. This makes it
possible to fix some of the accidental overlapping source inclusion in
modules within the User.gwt.xml cluster.

Patched the module loading process to pipe seen build and public
resources into the current libraryWriter.

Extended the moduleURL resolution system (depends on ResourceLoader) to
also find modules within the supplied libraryGroup.

Turned off BinaryTypeReferenceRestrictionsChecker when compiling
libraries (some of the resolved references *will* be to bytecode from
dependency libraries).

Patched CompilationState's assimilateUnits() process to store validated
compilation units in the libraryWriter.

Added assertions to ensure consistency of path structure when
initializing ZipFileResource (so that its getLocation() result will
always be a valid URL), and fixed two nonstandard uses.

Change-Id: I5efdbccd80df58657c958037e17600e8c381c28e