Update patch set 1
Patch Set 1:

Discussed offline with John. We probably need to go with alternative.

Quoting Ray from an earlier discussion:

"GWT cannot implement isInstance(), isAssignableTo()/isAssignableFrom()
because doing so would require the compiler to store a lot more
runtime type information. Right now, the compiler will only record
such type information if it sees a cast or an instanceof with a
literal class name, e.g.

(Bar)Foo -> an attempt to cast to Bar, so every implementer of Bar
should record if it is an implementor
x instanceof Bar -> an attempt to instanceof Bar, same idea as above

Unlike the language syntax operators,
Class.isInstance/isAssignable*/etc work on runtime instances, so if
the compiler sees Class.isInstance(object) the left hand size could
potentially be any class, and the right hand side could potentially be
any literal. This would force the compiler to retain a much larger
table of type hierarchy information in the runtime JS, so we don't
support it for the same reasons we don't support java.lang.reflect."

Patch-set: 1