Update patch set 1
Patch Set 1: Code-Review+2

It could possibly be easy to fix the test with the xsiframe linker. in c.g.g.core.CrossSiteIframeLinker.gwt.xml, the AsyncFragmentLoader.LoadingStrategy is redefined to ScriptTagLoadingStrategy when the xsiframe linker is used, instead of the default XhrLoadingStrategy. So in the test, the "logging" strategy would have to be a LoggingScriptTagLoadingStrategy instead of a LoggingXhrLoadingStrategy.

We can do it later though: first fix the test, then modify it (or possibly duplicate it) to use the xsiframe linker.

Patch-set: 1
Label: Code-Review=+2