Update patch set 1
Patch Set 1: Verified+1

Verified on Windows.

When navigating the sources on localhost:9876, with GWT 2.6.0 the list of files in each folder is empty. With master+this patch, files correctly appear.

That means the first change in SourceMap is not strictly needed, only the second is (as reported in the issue tracker). I suppose this is because browsers automatically translate backslashes to forward slashes in links.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=+1