Update patch set 1
Patch Set 1:

> What about just using File.seperator instead of '/'?

We're actually dealing with paths that will end up in URLs, so we need to use forward slashes and/or we get forward slashes in the input, as it comes from the requested URL.
java.util.File here was just an easy way to split the filename from the path. It works great on platforms where File.separator is "/", but on Windows java.util.File#getParent() will have normalized to backslashes which we cannot compare back to our inputs.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0