Update patch set 2
Patch Set 2:

> I thought Date object was already doing that. I actually verified
 > that while working on https://gwt-review.googlesource.com/#/c/4630/.
 > 
 > I'm kind of confused.

It is a different issue, and it was a change in Firefox that triggered the bug.

The issue is that new Date(109, 9, 31, 1, 30) is at first sight ambiguous.  It might refer to 10/31/2009 1:30 PST or 10/31/2009 1:30 PDT. Java returns the latter date (PST, as this is in a transition PDT->PST). We used to rely on the browser to solve this ambiguity and it seems that Firefox changed its behaviour and returns PDT.

Patch-set: 2
Label: Verified=0