Generate JavaScript code for super dispatch for modular compilation.
Dispatch through super.method(...) are effectively static and the
MakeCallsStatic pass immediately replaces them by their static
versions; but MakeCallsStatic statifies all other monomorphic calls
which is good in optimized builds but detrimental to separate
compilation as no function is ever pruned.

Alternatively we implement super dispatch by retrieving the method
from the corresponding prototype and calling with .call(....), like
we were doing for this(...) and super(...) constructor calls. The
difference between constructors and instance methods is that the former
have top level names while the latter have not.

If the MakeCallsStatic pass is called, only dispatches to super(...)
and this(...) constructors but not super.method(...) arrive at
GenerateJavaScriptAST hence leaving the optimized code path unaltered.

Change-Id: Ib02b23949c93ce6c2540e1ac98b5656d7a5a277d