Update patch set 3
Patch Set 3:

That field was added in a previous refactor of cast handling (https://gwt-review.googlesource.com/#/c/6176). You are right that elementTypeCategory is also used for what SEED was used before. But the main purpose of this field is to store information that was previously encoded in queryId (where negative queryId enocded java_object_or_jso and -1 encoded specifically JSO) and is what is used in setCheck().

The elementType field does not use the sign to encode other properties  (as it is a string in modular compile) and only encodes the "java" type of the element; it is elementTypeCategory that encodes whether it accepts JSOs, etc. By adding more information in this field now makes what was called SEED before unnecessary.

So this patch only optimizes casts/instanceof checks.

In this patch it is only renamed from elementTypeClass to elementTypeCategory.

Patch-set: 3